+++ src/lib/iv.integration-errors.spec.ts
@@ -5,7 +5,7 @@ describe('testing for errors', () => {
   let iv: IV;
   beforeEach(() => iv = new IV())
 
-  test('does not hang when returning to a node which previously ended with "goto"', async () => {
+  test('does not hang when returning to a node which previously ended with "goToNode"', async () => {
     /*
       There was once the concept in a runner of `shouldRun`. This state not getting properly reset
       caused this error. This was fixed. Subsequently, `shouldRun` was removed in favor of inferring
@@ -14,7 +14,7 @@ describe('testing for errors', () => {
     iv.variables = { count: 0 }
     iv.node('anything')
       .calculate({ storeIn: 'count', add: 1, var: 'count' })
-      .goto('anything else')
+      .goToNode('anything else')
 
     iv.node('anything else')
       .calculate({ storeIn: 'count', add: 1, var: 'count' })
@@ -28,7 +28,7 @@ describe('testing for errors', () => {
     expect(iv.variables.count).toBeGreaterThan(2)
   });
 
-  test('does not hang when goto is too quick', async () => {
+  test('does not hang when goToNode is too quick', async () => {
     /*
       This error came about because of a race condition. The former node had not yet finished
       executing when the run method was called again. This effectively did nothing.
@@ -37,13 +37,13 @@ describe('testing for errors', () => {
     */
     iv.variables = { count: 0 }
     iv.node('first')
-      .goto('second')
+      .goToNode('second')
 
     iv.node('second')
       .calculate({ storeIn: 'count', add: 1, var: 'count' })
       .if({var: 'count', isLessThan: 5})
         // .setVariable({storeIn: 'meh', value: 'meh'})  // This line makes the test pass. It should pass without it.
-        .goto('first')
+        .goToNode('first')
       .else()
         .setVariable({storeIn: 'worked', value: 'true'})
       .endIf()